Move cors setting
